rename: 패키지 이름에 언더바를 사용하지 않도록 변경
