--- src/main/java/roomescape/global_exception/BadRequestException.java
@@ -1,10 +1,0 @@
-package roomescape.global_exception;
-
-import org.springframework.http.HttpStatus;
-
-public class BadRequestException extends CustomException {
-
-    public BadRequestException(String message) {
-        super(HttpStatus.BAD_REQUEST, message);
-    }
-}
