--- src/main/java/roomescape/global_exception/ConflictException.java
@@ -1,10 +1,0 @@
-package roomescape.global_exception;
-
-import org.springframework.http.HttpStatus;
-
-public class ConflictException extends CustomException {
-
-    public ConflictException(String message) {
-        super(HttpStatus.CONFLICT, message);
-    }
-}
