--- src/main/java/roomescape/global_exception/CustomException.java
@@ -1,17 +1,0 @@
-package roomescape.global_exception;
-
-import org.springframework.http.HttpStatus;
-
-public abstract class CustomException extends RuntimeException {
-
-    private final HttpStatus status;
-
-    public CustomException(HttpStatus errorCode, String message) {
-        super(message);
-        this.status = errorCode;
-    }
-
-    public int getStatusValue() {
-        return status.value();
-    }
-}
