--- src/main/java/roomescape/global_exception/ErrorResponse.java
@@ -1,7 +1,0 @@
-package roomescape.global_exception;
-
-public record ErrorResponse(
-    String message
-) {
-}
-
