--- src/main/java/roomescape/global_exception/NotFoundException.java
@@ -1,10 +1,0 @@
-package roomescape.global_exception;
-
-import org.springframework.http.HttpStatus;
-
-public class NotFoundException extends CustomException {
-
-    public NotFoundException(String message) {
-        super(HttpStatus.NOT_FOUND, message);
-    }
-}
