+++ src/main/java/roomescape/globalexception/BadRequestException.java
@@ -1,0 +1,10 @@
+package roomescape.globalexception;
+
+import org.springframework.http.HttpStatus;
+
+public class BadRequestException extends CustomException {
+
+    public BadRequestException(String message) {
+        super(HttpStatus.BAD_REQUEST, message);
+    }
+}
