+++ src/main/java/roomescape/globalexception/ConflictException.java
@@ -1,0 +1,10 @@
+package roomescape.globalexception;
+
+import org.springframework.http.HttpStatus;
+
+public class ConflictException extends CustomException {
+
+    public ConflictException(String message) {
+        super(HttpStatus.CONFLICT, message);
+    }
+}
