+++ src/main/java/roomescape/globalexception/CustomException.java
@@ -1,0 +1,17 @@
+package roomescape.globalexception;
+
+import org.springframework.http.HttpStatus;
+
+public abstract class CustomException extends RuntimeException {
+
+    private final HttpStatus status;
+
+    public CustomException(HttpStatus errorCode, String message) {
+        super(message);
+        this.status = errorCode;
+    }
+
+    public int getStatusValue() {
+        return status.value();
+    }
+}
