+++ src/main/java/roomescape/globalexception/ErrorResponse.java
@@ -1,0 +1,7 @@
+package roomescape.globalexception;
+
+public record ErrorResponse(
+    String message
+) {
+}
+
