+++ src/main/java/roomescape/globalexception/NotFoundException.java
@@ -1,0 +1,10 @@
+package roomescape.globalexception;
+
+import org.springframework.http.HttpStatus;
+
+public class NotFoundException extends CustomException {
+
+    public NotFoundException(String message) {
+        super(HttpStatus.NOT_FOUND, message);
+    }
+}
