+++ src/main/java/roomescape/reservation/repository/ReservationRepositoryImpl.java
@@ -9,9 +9,9 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.global_exception.NotFoundException;
+import roomescape.globalexception.NotFoundException;
 import roomescape.reservation.domain.Reservation;
-import roomescape.reservation_time.domain.ReservationTime;
+import roomescape.reservationtime.domain.ReservationTime;
 import roomescape.theme.domain.Theme;
 
 @Repository
