+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -3,17 +3,17 @@ package roomescape.reservation.service;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
-import roomescape.global_exception.BadRequestException;
-import roomescape.global_exception.ConflictException;
+import roomescape.globalexception.BadRequestException;
+import roomescape.globalexception.ConflictException;
 import roomescape.reservation.ReservationMapper;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.dto.ReservationReqDto;
 import roomescape.reservation.domain.dto.ReservationResDto;
 import roomescape.reservation.repository.ReservationRepository;
-import roomescape.reservation_time.ReservationTimeMapper;
-import roomescape.reservation_time.domain.ReservationTime;
-import roomescape.reservation_time.domain.dto.ReservationTimeResDto;
-import roomescape.reservation_time.repository.ReservationTimeRepository;
+import roomescape.reservationtime.ReservationTimeMapper;
+import roomescape.reservationtime.domain.ReservationTime;
+import roomescape.reservationtime.domain.dto.ReservationTimeResDto;
+import roomescape.reservationtime.repository.ReservationTimeRepository;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.dto.ThemeResDto;
 import roomescape.theme.repository.ThemeRepository;
