--- src/main/java/roomescape/reservation_time/domain/dto/ReservationTimeResDto.java
@@ -1,6 +1,0 @@
-package roomescape.reservation_time.domain.dto;
-
-import java.time.LocalTime;
-
-public record ReservationTimeResDto(Long id, LocalTime startAt) {
-}
