--- src/main/java/roomescape/reservation_time/repository/ReservationTimeRepository.java
@@ -1,20 +1,0 @@
-package roomescape.reservation_time.repository;
-
-import java.util.List;
-import java.util.Optional;
-import roomescape.reservation_time.domain.ReservationTime;
-
-public interface ReservationTimeRepository {
-
-    List<ReservationTime> findAll();
-
-    ReservationTime findByIdOrThrow(Long id);
-
-    Optional<ReservationTime> findById(Long id);
-
-    ReservationTime add(ReservationTime reservationTime);
-
-    void delete(Long id);
-
-    Long insertWithKeyHolder(ReservationTime reservationTime);
-}
