--- src/main/java/roomescape/reservation_time/repository/ReservationTimeRepositoryImpl.java
@@ -1,85 +1,0 @@
-package roomescape.reservation_time.repository;
-
-import java.sql.PreparedStatement;
-import java.util.List;
-import java.util.Optional;
-import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Repository;
-import roomescape.global_exception.NotFoundException;
-import roomescape.reservation_time.domain.ReservationTime;
-
-@Repository
-public class ReservationTimeRepositoryImpl implements ReservationTimeRepository {
-
-    private final JdbcTemplate jdbcTemplate;
-
-    public ReservationTimeRepositoryImpl(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    @Override
-    public List<ReservationTime> findAll() {
-        return jdbcTemplate.query(
-            "select id, start_at from reservation_time",
-            (resultSet, rowNum) -> {
-                ReservationTime reservationTime = new ReservationTime(
-                    resultSet.getLong("id"),
-                    resultSet.getTime("start_at").toLocalTime()
-                );
-                return reservationTime;
-            });
-    }
-
-    @Override
-    public ReservationTime findByIdOrThrow(Long id) {
-        return findById(id)
-            .orElseThrow(() -> new NotFoundException("해당 예약시간 id가 존재하지 않습니다."));
-    }
-
-    @Override
-    public Optional<ReservationTime> findById(Long id) {
-        String sql = "select id, start_at from reservation_time where id = ?";
-
-        try {
-            return Optional.ofNullable(jdbcTemplate.queryForObject(sql,
-                (resultSet, rowNum) ->
-                    new ReservationTime(
-                        resultSet.getLong("id"),
-                        resultSet.getTime("start_at").toLocalTime()
-                    ), id));
-        } catch (EmptyResultDataAccessException e) {
-            return Optional.empty();
-        }
-    }
-
-    @Override
-    public ReservationTime add(ReservationTime reservationTime) {
-        Long id = insertWithKeyHolder(reservationTime);
-        return findByIdOrThrow(id);
-    }
-
-    @Override
-    public void delete(Long id) {
-        String sql = "delete from reservation_time where id = ?";
-        jdbcTemplate.update(sql, id);
-    }
-
-    @Override
-    public Long insertWithKeyHolder(ReservationTime reservationTime) {
-        String sql = "INSERT INTO reservation_time (start_at) VALUES (?)";
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-
-        jdbcTemplate.update(connection -> {
-            PreparedStatement ps = connection.prepareStatement(
-                sql,
-                new String[]{"id"});
-            ps.setString(1, reservationTime.getStartAt().toString());
-            return ps;
-        }, keyHolder);
-
-        return keyHolder.getKey().longValue();
-    }
-}
