+++ src/main/java/roomescape/reservationtime/domain/dto/ReservationTimeResDto.java
@@ -1,0 +1,6 @@
+package roomescape.reservationtime.domain.dto;
+
+import java.time.LocalTime;
+
+public record ReservationTimeResDto(Long id, LocalTime startAt) {
+}
