+++ src/main/java/roomescape/theme/repository/ThemeRepositoryImpl.java
@@ -10,7 +10,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.global_exception.NotFoundException;
+import roomescape.globalexception.NotFoundException;
 import roomescape.theme.domain.Theme;
 
 @Repository
