--- src/test/java/roomescape/reservation_time/controller/ReservationTimeControllerTest.java
@@ -1,139 +1,0 @@
-package roomescape.reservation_time.controller;
-
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-
-import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
-import java.time.LocalTime;
-import java.util.List;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTestContextBootstrapper;
-import org.springframework.http.HttpStatus;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.BootstrapWith;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import roomescape.reservation_time.ReservationTimeTestDataConfig;
-import roomescape.reservation_time.domain.dto.ReservationTimeReqDto;
-import roomescape.reservation_time.domain.dto.ReservationTimeResDto;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT, classes = {
-    ReservationTimeTestDataConfig.class})
-@BootstrapWith(SpringBootTestContextBootstrapper.class)
-@ExtendWith({SpringExtension.class})
-@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
-public class ReservationTimeControllerTest {
-
-    @Autowired
-    private JdbcTemplate jdbcTemplate;
-
-    @Nested
-    @DisplayName("GET /times 요청")
-    class findAll {
-
-        @DisplayName("데이터가 있을 때 200 OK와 함께 예약 시간을 반환한다")
-        @Test
-        void findAll_success_whenDataExists() {
-            // given 
-            jdbcTemplate.update("INSERT INTO reservation_time (start_at) VALUES (?)", "15:40");
-
-            // when & then
-            List<ReservationTimeResDto> resDtos = RestAssured.given().log().all()
-                .when().get("/times")
-                .then().log().all()
-                .statusCode(HttpStatus.OK.value()).extract()
-                .jsonPath().getList(".", ReservationTimeResDto.class);
-
-            Integer count = jdbcTemplate.queryForObject("SELECT count(1) from reservation_time", Integer.class);
-            assertThat(resDtos.size()).isEqualTo(count);
-        }
-
-        @DisplayName("데이터가 없을 때도 200 OK와 빈 리스트를 반환한다")
-        @Test
-        void findAll_success_whenNoData() {
-            // when & then
-            List<ReservationTimeResDto> resDtos = RestAssured.given().log().all()
-                .when().get("/times")
-                .then().log().all()
-                .statusCode(HttpStatus.OK.value()).extract()
-                .jsonPath().getList(".", ReservationTimeResDto.class);
-
-            Integer count = jdbcTemplate.queryForObject("SELECT count(1) from reservation_time", Integer.class);
-            assertThat(resDtos.size()).isEqualTo(count);
-        }
-    }
-
-    @Nested
-    @DisplayName("POST /times 요청")
-    class add {
-
-        @DisplayName("유효한 입력 시 201 Created와 함께 예약 시간이 추가된다")
-        @Test
-        void add_success_whenValidInput() {
-            // given
-            LocalTime dummyTime = LocalTime.of(11, 22);
-
-            // when
-            ReservationTimeReqDto dto = new ReservationTimeReqDto(dummyTime);
-
-            // then
-            RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(dto)
-                .when().post("/times")
-                .then().log().all()
-                .statusCode(HttpStatus.CREATED.value());
-        }
-
-        @DisplayName("중복된 예약 시간 입력 시 409 Conflict를 반환한다")
-        @Test
-        void add_failure_whenDuplicateInput() {
-            // given
-            LocalTime dummyTime = LocalTime.of(11, 22);
-            jdbcTemplate.update("INSERT INTO reservation_time (start_at) VALUES (?)", dummyTime.toString());
-
-            // when
-            ReservationTimeReqDto dto = new ReservationTimeReqDto(dummyTime);
-
-            // then
-            RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(dto)
-                .when().post("/times")
-                .then().log().all()
-                .statusCode(HttpStatus.CONFLICT.value());
-        }
-    }
-
-    @Nested
-    @DisplayName("DELETE /times/{id} 요청")
-    class delete {
-
-        @DisplayName("존재하는 ID로 삭제 요청 시 204 No Content를 반환한다")
-        @Test
-        void delete_success_withExistId() {
-            RestAssured.given().log().all()
-                .when().delete("/times/1")
-                .then().log().all()
-                .statusCode(HttpStatus.NO_CONTENT.value());
-
-            Integer countAfterDelete = jdbcTemplate.queryForObject("SELECT count(1) from reservation", Integer.class);
-            assertThat(countAfterDelete).isEqualTo(0);
-        }
-
-        @DisplayName("존재하지 않는 ID로 삭제 요청 시 400 Bad Request를 반환한다")
-        @Test
-        void delete_failure_byNonExistenceId() {
-            Long maxValue = Long.MAX_VALUE;
-            RestAssured.given().log().all()
-                .when().delete("/times/" + maxValue)
-                .then().log().all()
-                .statusCode(HttpStatus.NOT_FOUND.value());
-        }
-    }
-}
